deviceaction: retain explicit part type UUID when formatting
When applying a device format to a partition, ordinarily the
format will implicitly result in a part type UUID being set.
The set_flag/unset_flag methods can set an implicit UUID, as
can the parted_partition.system property assignment.

If the Partition class has an explicit part type UUID requested,
and previously set when the partition was first created, this
is liable to get lost during formatting setup. Thus need to
re-apply the part type UUID when formatting.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>